% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Title compare national data with Comtrade data}
\usage{
compare(datadf, datalist)
}
\arguments{
\item{datalist}{output of extract() function}

\item{data}{national datasets}
}
\value{
results of comparison dataframe
}
\description{
Title compare national data with Comtrade data
}
\examples{
\dontrun{
dlist <- extract(year = "2018", rcode = "EGY", pcode = "ARE", ccode = "190230")
datadf <- ndata
comparedf <- compare(ndata, dlist)
}
}
