% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Get UN Comtrade data}
\usage{
extract(year, rcode, pcode, ccode)
}
\arguments{
\item{year}{Data for which year(s) of interest, as a character vector.}

\item{rcode}{Country(s) of interest(ISO code), as a character vector.("All" to represent all countries)}

\item{pcode}{Country(s) that have interacted with the reporter country(s). ("All" to represent all countries)}

\item{ccode}{commodity code, as a character vector}
}
\value{
a list of two data frames, country of interest is reporter and country of interest is partner
}
\description{
Get UN Comtrade data
}
\examples{
\dontrun{
dlist <- extract(year = "2018", rcode = "EGY", pcode = "ARE", ccode = "190230")
dlist <- extract(year = "2018,2017", rcode = "EGY", pcode = "ARE", ccode = "190230")
dlist <- extract(year = "2018,2017", rcode = "EGY", pcode = "ARE", ccode = c("190230", "190190"))
dlist <- extract(year = "2018", rcode = "EGY", pcode = c("ARE","CHN"), ccode = c("190230", "190190"))
dlist$`reporter=reporter`; dlist$`reporter=partner`
}
}
