% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getotal.R
\name{getotal}
\alias{getotal}
\title{Title extract data for relative calculating (total flow of exports/imports for that reporter and year; total flow of exports/imports for that reporter and year for that commodity)}
\usage{
getotal(rcode, year, ccode)
}
\arguments{
\item{rcode}{Country(s) of interest(ISO code), as a character vector.("All" to represent all countries)}

\item{year}{Data for which year(s) of interest, as a character vector.}

\item{cccode}{commodity code, as a character vector.}
}
\value{
A dataframe that stores total value for later relative calculating
}
\description{
Title extract data for relative calculating (total flow of exports/imports for that reporter and year; total flow of exports/imports for that reporter and year for that commodity)
}
\examples{
\dontrun{
for specific commodity
total_for_c <- getotal(rcode = "EGY", year = "2018", ccode= "190230")
rcode = "CHN", year = "2016", ccode = "2848"
for all commodities
total_for_all <- getotal(rcode, year, "all")
}
}
