% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulatedsp.R
\name{calculatedsp}
\alias{calculatedsp}
\title{Title Calulate discrepancy}
\usage{
calculatedsp(dlist)
}
\arguments{
\item{dlist}{list of two dataframes (country of interest is reporter and country of interest is partner)}
}
\value{
dataframe of discrepancy
}
\description{
Title Calulate discrepancy
}
\examples{
\dontrun{
dlist <- extract(year = "2018", rcode = "EGY", pcode = "ARE", ccode = "190230")
df_dsp <- calculatedsp(dlist)
}
}
